requirements.txt
playwright
python-telegram-bot
python-dotenv
requests
httpx
beautifulsoup4
